Struct strason::Json

pub struct Json(_);

A "stringly-typed" Json object. That is, either a value (represented as a String), or an object (represented as a map from Strings to Jsons).

Methods

impl Json

fn from_iter<I: Iterator<Item=Result<u8>>>(it: I) -> Result<Json, Error>

Construct a Json object by parsing a byte iterator, e.g. from a Reader

fn from_str(s: &str) -> Result<Json, Error>

Construct a Json object by parsing a string

fn from_reader<R: Read>(r: R) -> Result<Json, Error>

Construct a Json object from a reader

fn from_serialize<T: Serialize>(t: &T) -> Result<Json, Error>

Construct a Json object from a Serialize type

fn null(&self) -> Option<()>

Returns a null, if this is a null

fn bool(&self) -> Option<bool>

Returns the value, if this is a boolean

fn num(&self) -> Option<&str>

Returns the value, if this is a number

fn string(&self) -> Option<&str>

Returns the value, if this is a string

fn array(&self) -> Option<&[Json]>

Returns the value, if this is an array

fn object(&self) -> Option<&[(String, Json)]>

Returns the value, if this is an object

fn get(&self, index: &str) -> Option<&Json>

Obtain a reference to a specified member, if this is an object

fn len(&self) -> usize

Return the number of subobjects this object represents (so a count for Arrays and Objects). NOT a string length.

fn is_empty(&self) -> bool

Return whether the object is empty, if it is a collection

fn write_to<W: Write>(&self, w: W) -> Result<()>

Reserialize the object into a writer

fn to_bytes(&self) -> Vec<u8>

Reserialize the object to byte array

fn into_deserialize<T: Deserialize>(self) -> Result<T, Error>

Convert the Json object to something deserializable